+++ Implement Nested Queries.sql
@@ -1,5 +1,5 @@
-CREATE TABLE ps6161(id INT(10), name VARCHAR(20));
-CREATE TABLE vs6161(id INT(10), marks INT(10));
+CREATE TABLE ps6161(id NUMBER(10), name VARCHAR(20));
+CREATE TABLE vs6161(id NUMBER(10), marks NUMBER(10));
 
 SELECT a.id, a.name, b.marks
 FROM ps6161 a, vs6161 b
